[cmake] Fix --only-test handling
Handling of --only-test was patchy, and expected it to be a directory. This caused problems when trying to reproduce a SingleSource test.

Now, we parse the path and assert that it either points to a directory or one level under a directory (where the basename is expected to be the test name, so is passed through to make).