[ui] Add special rendering for the "producer" run property
If a test run sets the run info property "producer", display it in the run comparison page in the top table.

It doesn't matter what this property is set to; it's intended to be used to indicate the "provenance" of a run - what ran it and why - like a buildbot build URL or jenkins URL.

Altough there's no required format, add special handling or strings that look like buildbot URLs or URLs in general.